+++ src/main/gen/com/javampire/openscad/highlighting/OpenSCADHighlightingLexer.java
@@ -914,7 +914,7 @@ class OpenSCADHighlightingLexer implements FlexLexer {
   /**
    * Refills the input buffer.
    *
-   * @return      <code>false</code>, iff there was new input.
+   * @return      {@code false}, iff there was new input.
    *
    * @exception   java.io.IOException  if any I/O-Error occurs
    */
@@ -950,7 +950,7 @@ class OpenSCADHighlightingLexer implements FlexLexer {
 
 
   /**
-   * Returns the character at position <tt>pos</tt> from the
+   * Returns the character at position {@code pos} from the
    * matched text.
    *
    * It is equivalent to yytext().charAt(pos), but faster
@@ -974,7 +974,7 @@ class OpenSCADHighlightingLexer implements FlexLexer {
 
 
   /**
-   * Reports an error that occured while scanning.
+   * Reports an error that occurred while scanning.
    *
    * In a wellformed scanner (no or only correct usage of
    * yypushback(int) and a match-all fallback rule) this method
@@ -1055,11 +1055,11 @@ class OpenSCADHighlightingLexer implements FlexLexer {
       if (zzMarkedPosL > zzStartRead) {
         switch (zzBufferL.charAt(zzMarkedPosL-1)) {
         case '\n':
-        case '\u000B':  // fall though
-        case '\u000C':  // fall though
-        case '\u0085':  // fall though
-        case '\u2028':  // fall though
-        case '\u2029':  // fall though
+        case '\u000B':  // fall through
+        case '\u000C':  // fall through
+        case '\u0085':  // fall through
+        case '\u2028':  // fall through
+        case '\u2029':  // fall through
           zzAtBOL = true;
           break;
         case '\r':