+++ src/main/gen/com/javampire/openscad/psi/impl/OpenSCADQualificationExprImpl.java
@@ -32,10 +32,12 @@ public class OpenSCADQualificationExprImpl extends OpenSCADResolvableElementImpl
     return findNotNullChildByClass(OpenSCADExpr.class);
   }
 
+  @Override
   public ItemPresentation getPresentation() {
     return OpenSCADPsiImplUtil.getPresentation(this);
   }
 
+  @Override
   public PsiElement getNameIdentifier() {
     return OpenSCADPsiImplUtil.getNameIdentifier(this);
   }