+++ src/main/java/com/javampire/openscad/parser/OpenSCADParserDefinition.java
@@ -156,7 +156,7 @@ public class OpenSCADParserDefinition implements ParserDefinition {
         return new OpenSCADFile(viewProvider);
     }
 
-    public SpaceRequirements spaceExistanceTypeBetweenTokens(ASTNode left, ASTNode right) {
+    public SpaceRequirements spaceExistenceTypeBetweenTokens(ASTNode left, ASTNode right) {
         return SpaceRequirements.MAY;
     }